+++ app/src/main/java/org/dhis2/uicomponents/map/layer/types/SatelliteMapLayer.kt
@@ -1,20 +1,17 @@
-package org.dhis2.uicomponents.map.layer
+package org.dhis2.uicomponents.map.layer.types
 
 import com.mapbox.mapboxsdk.maps.MapboxMap
 import com.mapbox.mapboxsdk.maps.Style
+import org.dhis2.uicomponents.map.layer.MapLayer
 
 class SatelliteMapLayer(
     private val mapboxMap: MapboxMap,
     private val styleChangeCallback: (() -> Unit)?
-): MapLayer {
+) : MapLayer {
 
-    var satelliteStyleLoaded = false
     override fun showLayer() {
-        if(!satelliteStyleLoaded) {
-            mapboxMap.setStyle(Style.SATELLITE_STREETS) {
-                satelliteStyleLoaded = true
-                styleChangeCallback?.invoke()
-            }
+        mapboxMap.setStyle(Style.SATELLITE_STREETS) {
+            styleChangeCallback?.invoke()
         }
     }
